Backend/requirements/base.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
billiard==4.2.1
black==25.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.14
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.9.2
cryptography==46.0.2
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.12.0
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
google-auth==2.41.1
google-genai==1.44.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
identify==2.6.12
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
jiter==0.10.0
kombu==5.5.4
mypy_extensions==1.1.0
nodeenv==1.9.1
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
PyMuPDF==1.26.3
pyparsing==3.2.3
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
